requirements.txt
feedparser==6.0.10
openai==2.30.0
resend==0.8.0
jinja2==3.1.2
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
thefuzz==0.19.0
beautifulsoup4==4.12.2